mysite/requirements.txt
Django==2.1.1
django-taggit==0.23.0
Markdown==3.0.1
psycopg2==2.7.5
